refactor: adjusted server structure
