+++ clear-ledger-server/src/main/java/com/onixbyte/clearledger/common/Formatters.java
@@ -1,4 +1,4 @@
-package com.onixbyte.clearledger.utils;
+package com.onixbyte.clearledger.common;
 
 import java.time.format.DateTimeFormatter;
