+++ clear-ledger-server/src/main/java/com/onixbyte/clearledger/config/ConcurrentConfig.java
@@ -1,6 +1,5 @@
 package com.onixbyte.clearledger.config;
 
-import com.onixbyte.clearledger.property.ConcurrentProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
