+++ clear-ledger-server/src/main/java/com/onixbyte/clearledger/config/CorsProperty.java
@@ -1,4 +1,4 @@
-package com.onixbyte.clearledger.property;
+package com.onixbyte.clearledger.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.web.bind.annotation.RequestMethod;
