+++ clear-ledger-server/src/main/java/com/onixbyte/clearledger/config/JacksonConfig.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.datatype.jsr310.deser.LocalTimeDeserializer;
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateSerializer;
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalTimeSerializer;
-import com.onixbyte.clearledger.utils.Formatters;
+import com.onixbyte.clearledger.common.Formatters;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
