+++ clear-ledger-server/src/main/java/com/onixbyte/clearledger/config/SecurityConfig.java
@@ -1,7 +1,7 @@
 package com.onixbyte.clearledger.config;
 
 import com.onixbyte.clearledger.filter.UserAuthenticationFilter;
-import com.onixbyte.clearledger.security.provider.UsernamePasswordAuthenticationProvider;
+import com.onixbyte.clearledger.security.UsernamePasswordAuthenticationProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
