+++ clear-ledger-server/src/main/java/com/onixbyte/clearledger/generator/LedgerIdCreator.java
@@ -1,8 +1,8 @@
-package com.onixbyte.clearledger.guid;
+package com.onixbyte.clearledger.generator;
 
 import com.onixbyte.clearledger.constant.IdType;
 import com.onixbyte.clearledger.service.SerialService;
-import com.onixbyte.clearledger.utils.Formatters;
+import com.onixbyte.clearledger.common.Formatters;
 import com.onixbyte.guid.GuidCreator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -10,18 +10,18 @@ import org.springframework.stereotype.Component;
 import java.time.LocalDate;
 
 @Component
-public class TransactionIdCreator implements GuidCreator<String> {
+public class LedgerIdCreator implements GuidCreator<String> {
 
     private final SerialService serialService;
 
     @Autowired
-    public TransactionIdCreator(SerialService serialService) {
+    public LedgerIdCreator(SerialService serialService) {
         this.serialService = serialService;
     }
 
     @Override
     public String nextId() {
         var date = LocalDate.now().format(Formatters.SHORTENED_DATE_FORMATTER);
-        return "%s%s%04d".formatted(IdType.TRANSACTION.getCode(), date, serialService.nextSerial("tx"));
+        return "%s%s%04d".formatted(IdType.LEDGER.getCode(), date, serialService.nextSerial("ledger"));
     }
 }
